+++ 3DDirectX1/others/Ease.cpp
@@ -92,7 +92,7 @@ float Eas::easeInCubic(float x)
 
 float Eas::easeOutCubic(float x)
 {
-	return 1 - pow(1 - x, 3);
+	return 1.0f - powf(1 - x, 3);
 }
 
 float Eas::easeInQuad(float x)
@@ -107,20 +107,20 @@ float Eas::easeOutQuad(float x)
 
 float Eas::easeOutBounce(float x)
 {
-	const float n1 = 7.5625;
-	const float d1 = 2.75;
+	const float n1 = 7.5625f;
+	const float d1 = 2.75f;
 
 	if (x < 1 / d1) {
 		return n1 * x * x;
 	}
 	else if (x < 2 / d1) {
-		return n1 * (x -= 1.5 / d1) * x + 0.75;
+		return n1 * (x -= 1.5f / d1) * x + 0.75f;
 	}
 	else if (x < 2.5 / d1) {
-		return n1 * (x -= 2.25 / d1) * x + 0.9375;
+		return n1 * (x -= 2.25f / d1) * x + 0.9375f;
 	}
 	else {
-		return n1 * (x -= 2.625 / d1) * x + 0.984375;
+		return n1 * (x -= 2.625f / d1) * x + 0.984375f;
 	}
 }
